pyproject.toml
[project]
name = "db-query-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.9.0",
    "langchain>=0.1.0",
    "langchain-ollama>=0.1.0",
    "python-dotenv>=1.0.0",
]
